Move main.dart into app/ per the doc; keep a thin shim at lib/main.dart
The architecture doc lists main.dart under app/, so the real bootstrap
now lives at lib/app/main.dart alongside app.dart, app_theme.dart, and
l10n/. lib/main.dart is now a 4-line shim that delegates to it, kept
only so `flutter run` and `flutter build` work without `-t` — Flutter's
tooling assumes lib/main.dart is the entry point.

No behavior change. Both files combined are still smaller than the
original main.dart was before the app/ extraction.
